+++ firebase-ai/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+- [feature] Added support for configuring the aspect ratio nad image size when generating images
+  with the Gemini Image models.
+- [feature] Added FinishReasons which can be returned from Gemini image generation.
+
 # 17.7.0
 
 - [changed] Added `LiveAudioConversationConfig` to control different aspects of the conversation